fix: refactor network status listener
