+++ .eslintignore
@@ -1,2 +1,3 @@
 # don't ever lint node_modules
 node_modules
+examples
