+++ examples/react/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
     "@types/react": "^16.8.0",
+    "apollo-cache-inmemory": "^1.6.5",
     "apollo-client": "^2.6.4",
     "apollo-link-ws": "^1.0.19",
     "graphql": "^14.3.1",
