+++ examples/server/.npmignore
@@ -1,0 +1,5 @@
+test/
+coverage/
+.nyc_output/
+node_modules/
+integration_test
